requirements.txt
dash
dash_bootstrap_components
dash_html_components
dash_core_components
Flask-SQLAlchemy
unidecode
Pillow
tweepy
numpy
pandas
matplotlib
pybase64
wordcloud
gunicorn
nltk
